requirements-dev.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.6
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.8.2
drf-yasg==1.8.0
future==0.16.0
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2018.4
requests==2.18.4
ruamel.yaml==0.15.37
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
